[FIX] Remove obsolete ir.model.fields and ir.model entries
While Odoo deletes obsolete field and model entries from the data model
metadata explicitely in their migration scripts, in OpenUpgrade I have
always meant to rely on the mechanism of purging 'untouched' XMLIDs that
takes care of the deletion of obsolete data records (e.g. views).

However, this mechanism was not applied to field and model entries because
their XMLIDs were created with noupdate NULL instead of FALSE and as such
excluded in the query to gather all obsolete data records.

Also missing was marking the XMLIDs of fields and models as `loaded` in the
first place.

All of this is working properly in Odoo 14 (introduced gradually across new
releases) so all of this is backported from newer versions one way or
another.