[FIX] use_parent_address <=> partner type 'contact'
In Odoo 10.0, the explicit flag `use_parent_address` to trigger the syncing of
address fields from the top level parent to their child partners is replaced
by logic depending on partner type: if partner type of the child partner is
contact, the address will be synced.

Therefore, switch child partners between type 'contact' and type 'order'
depending on the value of the obsolete `use_parent_address`.